Fix allProperties of theme editor
