+++ AuroraEditor/.githash
@@ -1,1 +1,1 @@
-6c949f8c608f3a87783ba80dfc2b5675eeb55443
+ad6ac0d34994d409de0052e07ba0cb3fee43e455
